Created script to collapse the data trees.
The script moves and renames all the data files scattered in the
directory trees found in the original backup drive (But I'd recommend
using rsync to copy the data trees first). New locations are organized by
type rather than in deep trees based on collection time. New names are
unique and use a more coherent format between the different types.

This script was intended to not be needed after it runs successfully
once (which it just did), but may be a nice reference for future work.

Changed gitignore to ignore the new log files.